move popup callbacks from editor to popups
